Use HTML table elements
