+++ web/src/components/datatable/body/body.module.scss
@@ -1,0 +1,3 @@
+.data-table {
+  width: 100%;
+}
