+++ web/src/components/datatable/body/index.ts
@@ -1,2 +1,2 @@
-export * from './column';
 export * from './body';
+export * from './row';
