+++ web/src/components/datatable/body/row.module.scss
@@ -1,0 +1,33 @@
+@import '@/scss/colors';
+
+.header-row,
+%row-base {
+  margin: 0;
+  font-size: 14px;
+  line-height: 20px;
+  font-weight: bold;
+  padding: 12px 16px;
+  text-align: left;
+
+  @include theme('table') {
+    background-color: t('bg-header-row');
+    color: t('text-header-row');
+  }
+
+  &-sortable {
+    @extend %row-base;
+
+    text-decoration: underline;
+    cursor: pointer;
+  }
+}
+
+.data-row {
+  padding: 12px 16px;
+
+  @include theme('table') {
+    border-bottom: 1px solid t('border-data-row');
+    background-color: t('bg-data-row');
+    color: t('text-data-row');
+  }
+}
