+++ web/src/scss/colors.scss
@@ -48,11 +48,11 @@ $themes: (
     ),
     'table': (
       'bg-header': $color-gray-800,
-      'bg-col': $color-gray-800,
-      'bg-col-header': $color-gray-700,
-      'col-text': $color-gray-400,
-      'col-text-header': $color-gray-400,
-      'cell-border': $color-gray-700,
+      'bg-header-row': $color-gray-700,
+      'text-header-row': $color-gray-400,
+      'bg-data-row': $color-gray-800,
+      'text-data-row': $color-gray-400,
+      'border-data-row': $color-gray-700,
     ),
   ),
   'light': (
@@ -69,11 +69,11 @@ $themes: (
     ),
     'table': (
       'bg-header': $color-white,
-      'bg-col': $color-white,
-      'bg-col-header': $color-gray-50,
-      'col-text': $color-gray-400,
-      'col-text-header': $color-gray-700,
-      'cell-border': $color-gray-200,
+      'bg-header-row': $color-gray-50,
+      'text-header-row': $color-gray-700,
+      'bg-data-row': $color-white,
+      'text-data-row': $color-gray-400,
+      'border-data-row': $color-gray-200,
     ),
   ),
 );
